footer + reviewPage axios import rename
