+++ src/components/Footer/Footer.module.css
@@ -1,3 +1,4 @@
 .footer {
-  background-color: #bfbb9c;
+  background-color: #565a5b;
+  border-radius: 2rem 2rem 0 0;
 }
