+++ src/components/Main/ReviewsPage/ReviewsPage.module.css
@@ -5,6 +5,7 @@
   border-radius: 2rem;
   box-shadow: 0 1px 9px 1px;
   background-color: #87897c;
+  margin-bottom: 1rem;
 }
 
 .reviewFilterAside {
